fix: remarks from reviewer
