+++ src/Visitors/AppBootstrapVisitors/AbstractAppBootstrapVisitor.php
@@ -3,11 +3,15 @@
 namespace RonasIT\Larabuilder\Visitors\AppBootstrapVisitors;
 
 use PhpParser\Node;
+use PhpParser\Node\Arg;
+use PhpParser\Node\Expr\Closure;
 use PhpParser\Node\Expr\MethodCall;
+use PhpParser\Node\Identifier;
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\Enum_;
 use PhpParser\Node\Stmt\Expression;
 use PhpParser\Node\Stmt\Interface_;
+use PhpParser\Node\Stmt\Nop;
 use PhpParser\Node\Stmt\Trait_;
 use RonasIT\Larabuilder\Exceptions\InvalidBootstrapAppFileException;
 use RonasIT\Larabuilder\Visitors\BaseNodeVisitorAbstract;
@@ -21,7 +25,7 @@ abstract class AbstractAppBootstrapVisitor extends BaseNodeVisitorAbstract
         Enum_::class,
     ];
 
-    abstract protected function insertNode(MethodCall $node): MethodCall;
+    abstract protected function getInsertableNode(): Expression;
 
     public function __construct(
         protected string $parentMethod,
@@ -40,10 +44,56 @@ abstract class AbstractAppBootstrapVisitor extends BaseNodeVisitorAbstract
 
     public function leaveNode(Node $node): Node
     {
+        if ($node instanceof MethodCall && $node->name->toString() === 'create') {
+            if ($this->isParentMethodMissing($node)) {
+                $node = $this->insertParentNode($node);
+            }
+
+            if ($node->var->getAttribute('wasCreated')) {
+                $node->var = $this->handleParentNode($node->var);
+            }
+        }
+
         if (!$node instanceof MethodCall) {
             return $node;
         }
 
+        if ($this->isParentNode($node) && $this->shouldInsertNode($node)) {
+            return $this->insertNode($node);
+        }
+
+        return $node;
+    }
+
+    protected function isParentMethodMissing(Node $node): bool
+    {
+        $nodeVar = $node->var;
+
+        while ($nodeVar instanceof MethodCall) {
+            if ($nodeVar->name->toString() === $this->parentMethod) {
+                return false;
+            }
+
+            $nodeVar = $nodeVar->var;
+        }
+
+        return true;
+    }
+
+    protected function insertParentNode(Node $node): Node
+    {
+        $closure = new Closure([
+            'returnType' => new Identifier('void'),
+        ]);
+
+        $parentCall = new MethodCall($node->var, new Identifier($this->parentMethod), [new Arg($closure)]);
+        $parentCall->setAttribute('wasCreated', true);
+
+        return new MethodCall($parentCall, new Identifier('create'));
+    }
+
+    protected function handleParentNode(MethodCall $node): Node
+    {
         if ($this->isParentNode($node) && $this->shouldInsertNode($node)) {
             return $this->insertNode($node);
         }
@@ -73,6 +123,26 @@ abstract class AbstractAppBootstrapVisitor extends BaseNodeVisitorAbstract
         }
 
         return true;
+    }
+
+    protected function insertNode(MethodCall $node): MethodCall
+    {
+        $currentStatements = $node->args[0]->value->stmts;
+        $statement = $this->getInsertableNode();
+
+        if (count($currentStatements) === 1 && $currentStatements[0] instanceof Nop) {
+            $node->args[0]->value->stmts = [$statement];
+
+            return $node;
+        }
+
+        $lastExistingStatement = end($currentStatements);
+
+        $statement->setAttribute('previous', $lastExistingStatement);
+
+        $node->args[0]->value->stmts[] = $statement;
+
+        return $node;
     }
 
     protected function matchesCustomCriteria(Expression $statement): bool
