+++ src/Visitors/AppBootstrapVisitors/AddExceptionsRender.php
@@ -10,7 +10,6 @@ use PhpParser\Node\Identifier;
 use PhpParser\Node\Name;
 use PhpParser\Node\Param;
 use PhpParser\Node\Stmt\Expression;
-use PhpParser\Node\Stmt\Nop;
 use RonasIT\Larabuilder\Nodes\PreformattedCode;
 
 class AddExceptionsRender extends AbstractAppBootstrapVisitor
@@ -28,38 +27,6 @@ class AddExceptionsRender extends AbstractAppBootstrapVisitor
             parentMethod: 'withExceptions',
             targetMethod: 'render',
         );
-    }
-
-    protected function matchesCustomCriteria(Expression $stmt): bool
-    {
-        $paramType = $stmt->expr->args[0]?->value?->params[0]?->type ?? null;
-
-        if (!($paramType instanceof Name)) {
-            return false;
-        }
-
-        $typeName = $paramType->toString();
-
-        return $typeName === $this->exceptionClass || $typeName === class_basename($this->exceptionClass);
-    }
-
-    protected function insertNode(MethodCall $node): MethodCall
-    {
-        $currentStatements = $node->args[0]->value->stmts;
-
-        if (count($currentStatements) === 1 && $currentStatements[0] instanceof Nop) {
-            $node->args[0]->value->stmts = [$this->renderStatement];
-
-            return $node;
-        }
-
-        $lastExistingStatement = end($currentStatements);
-
-        $this->renderStatement->setAttribute('previous', $lastExistingStatement);
-
-        $node->args[0]->value->stmts[] = $this->renderStatement;
-
-        return $node;
     }
 
     protected function buildRenderCall(): Expression
@@ -95,5 +62,23 @@ class AddExceptionsRender extends AbstractAppBootstrapVisitor
             'params' => $params,
             'stmts' => [new PreformattedCode($this->renderBody)],
         ]);
+    }
+
+    protected function matchesCustomCriteria(Expression $stmt): bool
+    {
+        $paramType = $stmt->expr->args[0]?->value?->params[0]?->type ?? null;
+
+        if (!($paramType instanceof Name)) {
+            return false;
+        }
+
+        $typeName = $paramType->toString();
+
+        return $typeName === $this->exceptionClass || $typeName === class_basename($this->exceptionClass);
+    }
+
+    protected function getInsertableNode(): Expression
+    {
+        return $this->renderStatement;
     }
 }
